requirements/base.txt
Django==4.2.5
python-decouple==3.8
dj-database-url==2.1.0
